% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth.R
\name{add_ENSID}
\alias{add_ENSID}
\title{add ENSID}
\usage{
add_ENSID(Kmeans_result, GeneInf1 = NULL, Spec1 = "")
}
\arguments{
\item{Kmeans_result}{Kmeans result data.frame, row names should be Symbol ID}

\item{GeneInf1}{data.frame, correspondence file of gene ID, row names should be ENSEMBLE ID, first column should be Symbol ID, if GeneInf1 = NULL this function will be built-in corresponding gene ID file.}

\item{Spec1}{If you don’t have a gene ID corresponding file, you can also use our built-in corresponding gene ID file, 'Mm' for mus musculus}
}
\value{
add ENSEMBL ID to the first column of data frame
}
\description{
Add ENSEMBLE ID to the first column of your data.frame
}
\examples{
add_ENSID(clustering, Spec1 = "Hs")
}
