% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Footprints_FOS.R
\name{Add_expression_information}
\alias{Add_expression_information}
\title{Internal function}
\usage{
Add_expression_information(Regulation1, GeneM0)
}
\arguments{
\item{Regulation1}{regulaotry relationship}

\item{GeneM0}{expression profile}
}
\value{
regulatory relationships with expression information
}
\description{
Internal function
}
