% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA1.R
\name{add_pseudotime}
\alias{add_pseudotime}
\title{add pseudotime in monocle object to the metadata of the seurat object}
\usage{
add_pseudotime(seurat_object, monocle_object)
}
\arguments{
\item{seurat_object}{seurat object, which has same cells and genes as monocle_object}

\item{monocle_object}{monocle object, which include pesudotime and has same
cells and genes as seurat_object}
}
\value{
seurat_object with pseudotime
}
\description{
add pseudotime in monocle object to the metadata of the seurat object
}
\examples{
load(system.file("extdata", "test_seurat.rda", package = "IReNA"))
monocle_object = get_pseudotime(test_seurat)
add_pseudotime_DEG_filter(seurat_object = test_seurat,monocle_object = monocle_object)
}
