% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA1.R
\name{add_pseudotime_DEG_filter}
\alias{add_pseudotime_DEG_filter}
\title{Add pseudotime to seurat object, and filter differential expressed genes according to pseudotime}
\usage{
add_pseudotime_DEG_filter(
  seurat_object,
  monocle_object,
  DEG = TRUE,
  qvalue = 0.001,
  nce = 0.1,
  ed = 0.1,
  normlize1 = TRUE
)
}
\arguments{
\item{seurat_object}{seurat object, which has same cells and genes as monocle_object}

\item{monocle_object}{monocle object, which include pesudotime and has same cells and genes as seurat_object}

\item{DEG}{TRUE or FALSE, whether filter differential expressed genes for seurat object, default is TRUE}

\item{qvalue}{q-value, one of the indicators for filtering differential genes, default is 0.001}

\item{nce}{num cells expressed, one of the indicators for filtering differential genes, default is 0.1}

\item{ed}{expression difference, one of the indicators for filtering differential genes, default is 0.1}

\item{normlize1}{TRUE or FALSE, whether normalize the data in seurat object}
}
\value{
return a seurat object
}
\description{
Add pseudotime to seurat object, and filter differential expressed genes according to pseudotime
}
\examples{
monocle_object = get_pseudotime(seurat_object)
#add_pseudotime_DEG_filter(seurat_object = seurat_object,monocle_object = monocle_object, DEG = TRUE, qvalue = 0.001, nce = 0.1, ed = 0.1)
}
