% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA1.R
\name{add_regulation_type}
\alias{add_regulation_type}
\title{Add regulation type for each gene pair}
\usage{
add_regulation_type(Kmeans_result, potential_regulation, start_column = 4)
}
\arguments{
\item{Kmeans_result}{Kmeans result data.frame, row names should be ENSEMBEL ID,
and the first column should be gene Symbol ID, the second column should be KmeansGroup}

\item{potential_regulation}{First column is source gene, second column is target
gene, third column is weight. Several methods are recommended to generate
potential regulation: Genin3, PIDC, GRNBOOST2}

\item{start_column}{numeric, indicating the start column of expression value,
defalut is 3}
}
\value{

}
\description{
This function calculate the Spearman's correlation for each gene
pair as the regulation types. Correlation > 0 is positive regulation, Correlation
< 0 is negative regulation. Then, this function add the module and gene Symbol
for Kmeans_result to each gene in the gene pairs.
}
