% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Footprints_FOS.R
\name{Add_size_of_motif}
\alias{Add_size_of_motif}
\title{Inner function}
\usage{
Add_size_of_motif(cutsp_list, Candid)
}
\arguments{
\item{cutsp_list}{list that each element is cuptsp}

\item{Candid}{candidate gene}
}
\value{
return cutsp list with motif size
}
\description{
Inner function
}
