% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wig_track.R
\name{cal_footprint_cuts}
\alias{cal_footprint_cuts}
\title{Calculate cuts}
\usage{
cal_footprint_cuts(bamfilepath, bedfile, index_bam = FALSE, workers = NULL)
}
\arguments{
\item{bamfilepath}{character, indicating the path of bam file}

\item{bedfile}{bed file of footprints, generated by get_peaks_genes()}

\item{index_bam}{logical, indicating whether to index the bam file. If bam
file do not have a matched bai file, this parameter should be TRUE}

\item{workers}{number of cores}
}
\value{
return formated wig file
}
\description{
Function to calculate cuts of each position in footprints
}
\examples{
load(system.file("extdata", "list1.rda", package = "IReNA"))
load(system.file("extdata", "test_clustering.rda", package = "IReNA"))
bamfilepath1<-'mmATACCtrW00R1_CuFiQ10No_sorted.bam'
bamfilepath2<-'mmATACCtrW00R2_CuFiQ10No_sorted.bam'
test_clustering <- add_ENSID(test_clustering, Spec1 = "Hs")
list2<-get_related_peaks(list1,test_clustering)
#cuts1<-cal_footprint_cuts(bamfilepath = bamfilepath1,bedfile = list2[[1]])
#cuts2<-cal_footprint_cuts(bamfilepath = bamfilepath2,bedfile = list2[[1]])
#cuts_list<-list(cuts1,cuts2)
}
