% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Footprints_FOS.R
\name{Cal_Footprints_FOS}
\alias{Cal_Footprints_FOS}
\title{Inner function}
\usage{
Cal_Footprints_FOS(cutsp2_list, FlankFold1 = 3)
}
\arguments{
\item{cutsp2_list}{list that each element is cutsp2}

\item{FlankFold1}{threshold}
}
\value{
return FOS
}
\description{
Inner function
}
