% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Footprints_FOS.R
\name{Cal_Footprints_FOS2}
\alias{Cal_Footprints_FOS2}
\title{Inner function to calculate footprint occupancy score}
\usage{
Cal_Footprints_FOS2(FP1, FlankFold1 = 3)
}
\arguments{
\item{FP1}{data source}

\item{FlankFold1}{threshold}
}
\description{
Inner function to calculate footprint occupancy score
}
