% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Footprints_FOS.R
\name{Cal_GenePairs_Cor2}
\alias{Cal_GenePairs_Cor2}
\title{Inner function}
\usage{
Cal_GenePairs_Cor2(RNA1, Pair1)
}
\arguments{
\item{RNA1}{expression}

\item{Pair1}{pair}
}
\description{
Inner function
}
