% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Footprints_FOS.R
\name{Cal_Regulation_Cor}
\alias{Cal_Regulation_Cor}
\title{Internal function}
\usage{
Cal_Regulation_Cor(Regulation1, ScExp1, CorThr1 = 0.8, ScCorThr1 = 0.2)
}
\arguments{
\item{Regulation1}{regulatory relationship}

\item{ScExp1}{expression profile}

\item{CorThr1}{threshold}

\item{ScCorThr1}{threshold}
}
\description{
Internal function
}
