% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{ChscRNA_genes}
\alias{ChscRNA_genes}
\title{chicken Gene ID correspondence}
\format{
contain ENSEMBEL ID, Symbol ID and NCBI ID
}
\usage{
ChscRNA_genes
}
\description{
chicken Gene ID correspondence
}
\keyword{datasets}
