% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{clustering}
\alias{clustering}
\title{data for demonstrating}
\format{
expression profile, first column is ENSEMBLE ID, second column is Symbol
}
\usage{
clustering
}
\description{
data for demonstrating
}
\keyword{datasets}
