% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth.R
\name{clustering_Kmeans}
\alias{clustering_Kmeans}
\title{clustering DEGs by K-means}
\usage{
clustering_Kmeans(
  RNA1,
  K1 = 1,
  ColumnGroup1 = NULL,
  Scale1 = "row",
  Range1 = c(-Inf, Inf),
  Reorder1 = TRUE,
  RevOrder1 = -1,
  NAcolum1 = NULL
)
}
\arguments{
\item{RNA1}{expression profile}

\item{K1}{numbers of bin}

\item{ColumnGroup1}{vector, if it is not NULL, this function wil sepate columns
according to varible ColumnGroup1, default is NULL}

\item{Scale1}{character, indicating if the values should be centered and scaled
in either the row direction or the column direction, or none. Corresponding values are "row", "column" and "none"}

\item{Range1}{numeric, indicating the range of expression values, defalut is c(-Inf, Inf)}

\item{Reorder1}{logic, indicating whether to order gene through hclustering,
defalut is TRUE}

\item{RevOrder1}{logic, indicating whether to reverser the order}

\item{NAcolum1}{vector, indicating which columns need to be removed due to NA values}
}
\value{
return a data.frame, and the first column is K-means group
}
\description{
This function is used to cluster the DEGs in the expression
profile by K-means algorithmn
}
\examples{
load(system.file("extdata", "test_clustering.rda", package = "IReNA"))
expression_profile = test_clustering[,-1];expression_profile<-na.omit(expression_profile)
clustering_Kmeans(expression_profile, K1=4, Range1=c(-3,3))
}
