% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Footprints_FOS.R
\name{Combine_Footprints_FOS}
\alias{Combine_Footprints_FOS}
\title{Inner function}
\usage{
Combine_Footprints_FOS(FOS_list)
}
\arguments{
\item{FOS_list}{list contain footprints with FOS}
}
\description{
Inner function
}
