% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{diff_peaks}
\alias{diff_peaks}
\title{Identify differential peaks}
\usage{
diff_peaks(Count, Group)
}
\arguments{
\item{Count}{data.frame, indicating counts of peaks, generated by htseq-count. If you calculate counts for each sample separately, you need to run merge_sort_count() function firstly.}

\item{Group}{vector or factor giving the experimental group/condition for each sample/library.}
}
\value{
return FDR q-value and foldchange of each peak, which can be used to identify differential peaks.
}
\description{
Identify differential peaks based on edgeR, samples of peaks need to be divided into at least two groups.
}
\examples{
load(system.file("extdata", "test_peak.rda", package = "IReNA"))
group <- c(1,1,2,2)
peaks <- diff_peaks(test_peak[,4:7],group)
}
