% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct_network.R
\name{enrich_module}
\alias{enrich_module}
\title{Enrichment analysis of each module}
\usage{
enrich_module(
  Kmeans_result,
  org.db,
  enrich.db,
  fun_num = 5,
  pvalueCutoff = 0.05,
  use_internal_data = TRUE,
  organism = NULL
)
}
\arguments{
\item{Kmeans_result}{Kmeans result data.frame, row names should be ENSEMBEL ID,
and the first column should be gene Symbol ID, the second column should be KmeansGroup}

\item{org.db}{org.db of your species.}

\item{enrich.db}{'GO' for GO enricment analysis, 'KEGG' for KEGG enrichment analysis}

\item{fun_num}{numeric, indicating the number of output functions per module}

\item{pvalueCutoff}{adjusted pvalue cutoff on enrichment tests to report}

\item{use_internal_data}{logical, use KEGG.db or latest online KEGG data}

\item{organism}{character, used for KEGG enrichment analysis.
supported organism listed in 'http://www.genome.jp/kegg/catalog/org_list.html'}
}
\value{
data.frame contains enrichment functions for each module
}
\description{
This function integrate clusterprofile make enrichment analysis
for each moudle. Before run this function, you need to download the org.db for
your species through BiocManager.
}
\examples{
\dontrun{
library(org.Hs.eg.db)
Kmeans_cluster_Ens <- add_ENSID(clustering,Spec1='Hs')
enrichment <- enrich_module(Kmeans_cluster_Ens, org.Hs.eg.db, 'KEGG')
enrichment <- enrich_module(Kmeans_cluster_Ens, org.Hs.eg.db, 'GO')
}
}
