% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA1.R
\name{extract_expressed_TFs}
\alias{extract_expressed_TFs}
\title{extract expressed transcription factors}
\usage{
extract_expressed_TFs(seurat_object, TFs, cells_quantile = 0.05)
}
\arguments{
\item{seurat_object}{seurat object}

\item{TFs}{character, indicating transcription factors names}

\item{cells_quantile}{numeric, indicating proportion of cells that
transcription factors express (threshold to get expressed transcription factors)}
}
\value{

}
\description{
extract expressed transcription factors according to the proportion of cells that
transcription factors express
}
