% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth.R
\name{fileter_expression_profile}
\alias{fileter_expression_profile}
\title{this function is used to filter the bins in the expression profile which value is zero}
\usage{
fileter_expression_profile(expression_profile, filterfc = TRUE, FC = 0.1)
}
\arguments{
\item{expression_profile}{expression profile}

\item{filterfc}{whether you want to filter genes based on logFC}

\item{FC}{logFC threshold}
}
\value{
return filtered expression profile
}
\description{
this function is used to filter the bins in the expression profile which value is zero
}
\examples{
expression_profile = Kmeans_clustering[,-1];expression_profile<-na.omit(expression_profile)
fileter_expression_profile(expression_profile)
}
