% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Footprints_FOS.R
\name{filter_ATAC}
\alias{filter_ATAC}
\title{filter regulatory relationships based on footprints with high FOS}
\usage{
filter_ATAC(FOS, regulary_relationships)
}
\arguments{
\item{FOS}{footprints information, generated in \link{Footprints_FOS}}

\item{regulary_relationships}{generated in \link{get_cor}}
}
\value{
overlapped regulatory relationships
}
\description{
overlap footprints information with regulatory relationships
}
