% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Footprints_FOS.R
\name{Filter_Footprints}
\alias{Filter_Footprints}
\title{Inner function}
\usage{
Filter_Footprints(FOS1, FOS_cutoff)
}
\arguments{
\item{FOS1}{footprints occupancy score}

\item{FOS_cutoff}{numeric, indicating the threshold of footprint occupancy score to identify significant footprints}
}
\value{
return filtered FOS
}
\description{
Inner function
}
