% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA1.R
\name{filter_original_regulation}
\alias{filter_original_regulation}
\title{filter original regulation}
\usage{
filter_original_regulation(
  potential_regulation,
  motif,
  tf_threshold = 500,
  target_threshold = 100
)
}
\arguments{
\item{potential_regulation}{data.frame consists of three columns, first column
contains source genes, second column contains target genes, third column contains
weight.}

\item{motif}{motif file, you can choose our bulit-in motif database of
'mus musculus', 'homo sapiens', 'zebrafish' and 'chicken' by 'motif = Tranfac201803_Mm_MotifTFsF',
'motif = Tranfac201803_Hs_MotifTFsF', 'motif = Tranfac201803_Zf_MotifTFsF',
'motif = Tranfac201803_Ch_MotifTFsF' respectively, or you can upload your own motif data base, but the formata use be the same as our built-in motif database.}

\item{tf_threshold}{the numbers of targets with highest weight of each
transcription factor. Indicating the threshold to filter regulation.}

\item{target_threshold}{numbers of top transcription factors for each target.
Indicating the threshold to filter regulation.}
}
\description{
This function is used to further filter regulation generated by
Genin3, PIDC, or other methods.
}
