% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct_network.R
\name{filter_regulation}
\alias{filter_regulation}
\title{Filtering regulatory relationship through RcisTarget}
\usage{
filter_regulation(correlation, Species, Rankingsdir)
}
\arguments{
\item{correlation}{regulatory relationship, generated by get_cor() function}

\item{Species}{character, indicating species of your data, and used to choose annotation file in RcisTarget. Because of the limitaion of RcisTarget, this function only support three species: 'Mm', 'Hs' and 'Fly'}

\item{Rankingspath}{path of Gene-motif rankings file, which can be downloaded in https://resources.aertslab.org/cistarget. For more details about motif ranking file, you can read https://bioconductor.riken.jp/packages/3.9/bioc/vignettes/RcisTarget/inst/doc/RcisTarget.html#gene-motif_rankings}
}
\value{
return filtered regulatory relationship
}
\description{
Filtering regulatory relationship through RcisTarget
}
