% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_fimo.R
\name{filter_regulation_fimo}
\alias{filter_regulation_fimo}
\title{Filter regulatory relationships according to binding motifs generated by fimo}
\usage{
filter_regulation_fimo(fimo_regulation, regulatory_relationships)
}
\arguments{
\item{fimo_regulation}{regulation database, you can download it from xxx, or
generate it by \code{\link{generate_fimo_regulation}}}

\item{regulatory_relationships}{unfiltered regulatory_relationships, generated by
\code{\link{get_cor}}}
}
\description{
If ATAC-seq data is not available, binding motifs of target genes
can be used to filter regulatory relationships.
}
