% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_fimo.R
\name{find_motifs_targetgenes}
\alias{find_motifs_targetgenes}
\title{Title}
\usage{
find_motifs_targetgenes(
  gene_tss,
  motif,
  refdir,
  fimodir,
  outputdir1,
  Motifdir,
  sequencedir = NULL,
  select_motif = T
)
}
\arguments{
\item{gene_tss}{TSS region of genes. first column should be gene, second column
should be chr, third column should be start, fourth column should be end.
generated by \code{\link{get_tss_region}}}

\item{motif}{motif file, you can choose our bulit-in motif database of
'mus musculus', 'homo sapiens', 'zebrafish' and 'chicken' by 'motif = Tranfac201803_Mm_MotifTFsF',
'motif = Tranfac201803_Hs_MotifTFsF', 'motif = Tranfac201803_Zf_MotifTFsF',
'motif = Tranfac201803_Ch_MotifTFsF' respectively, or you can upload your own
motif data base, but the formata use be the same as our built-in motif database.}

\item{refdir}{character, indicating the path of reference genome. Reference
genome can be download from https://hgdownload.soe.ucsc.edu/downloads.html}

\item{fimodir}{character, indicating path of fimo software,
if you have added fimo to the environment variable, just set this argument as 'fimo'}

\item{outputdir1}{character, indicating the output path of the shell scripts
and sequence of target genes tss regions.(function 'find_motifs_targetgenes'
will automatically generate two folders ('fasta' and 'fimo') in the path
'outputdir1', and store sequence of target genes tss regions in the 'fasta'
and shell scripts in the 'fimo')}

\item{Motifdir}{character, indicating the path of meme motif file}

\item{sequencedir}{character, indicating the path of sequence of target genes
tss regions. If it's NULL, this parameter will be paste0(outputdir1,'fasta/')}

\item{select_motif}{logic, indicating whether to select motifs whose related
transcription factors are in genes of gene_tss}
}
\value{

}
\description{
Title
}
