% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Footprints_FOS.R
\name{Footprints_FOS}
\alias{Footprints_FOS}
\title{Calculate the FOS(footprints occupancy score) to identify enriched transcriptions factor, and use these transcription factors to identify significant regulatory relationships}
\usage{
Footprints_FOS(
  Wig_list,
  Candid,
  expression_profile,
  FlankFold1 = 3,
  trans_wig = FALSE
)
}
\arguments{
\item{Wig_list}{list that contains every data.frame generated by wig_track() or dnase_wig_tracks_both2.py, if you use the python script to get wig_list, please set trans_wig parameter as TRUE.}

\item{Candid}{If you follow our our pipline, this parameter should be the second element of the list that generated by get_peaks_genes()}

\item{expression_profile}{dataframe, indicating expression profile of bulk RNA-seq or single-cell RNA-seq, rownames should be ENSEMBEL ID of genes.}

\item{FlankFold1}{numeric, indicating}

\item{trans_wig}{logcial, if you use dnase_wig_tracks_both2.py to get wig_list, please set trans_wig parameter as TRUE.}
}
\value{
return regulatory relationships of enriched transcription fators to genes
}
\description{
Calculate the FOS(footprints occupancy score) to identify enriched transcriptions factor, and use these transcription factors to identify significant regulatory relationships
}
