% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_fimo.R
\name{generate_fimo_regulation}
\alias{generate_fimo_regulation}
\title{Generate regulation database accoding to the fimo result}
\usage{
generate_fimo_regulation(motif_dir, motif)
}
\arguments{
\item{motif_dir}{character, indicating the path of fimo result}

\item{motif}{motif file, you can choose our bulit-in motif database of
'mus musculus', 'homo sapiens', 'zebrafish' and 'chicken' by 'motif = Tranfac201803_Mm_MotifTFsF',
'motif = Tranfac201803_Hs_MotifTFsF', 'motif = Tranfac201803_Zf_MotifTFsF',
'motif = Tranfac201803_Ch_MotifTFsF' respectively, or you can upload your own
motif data base, but the formata use be the same as our built-in motif database.}
}
\value{

}
\description{
Generate regulation database accoding to the fimo result
}
