% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{generate_gtf}
\alias{generate_gtf}
\title{Merge peak and generate gft}
\usage{
generate_gtf(peak)
}
\arguments{
\item{peak}{data.frame, indicating peaks file. If you call peak for each sample separately, you need to combine these peaks firstly through cbind().}
}
\value{
return peaks related gtf file
}
\description{
Merge the peaks that are close to each other and modify the length of the peak to 500. Finally, return peaks related gtf file
}
\examples{
load(system.file("extdata", "test_peak.rda", package = "IReNA"))
peaks <- test_peak[,c(1,2,3)]
gtf <- generate_gtf(peaks)

\dontrun{
 sample1_peaks <- read.delim("sample1_peaks.narrowPeak", header=FALSE)
 sample2_peaks <- read.delim("sample2_peaks.narrowPeak", header=FALSE)
 sample3_peaks <- read.delim("sample3_peaks.narrowPeak", header=FALSE)
 sample4_peaks <- read.delim("sample4_peaks.narrowPeak", header=FALSE)
 sample5_peaks <- read.delim("sample5_peaks.narrowPeak", header=FALSE)
 sample6_peaks <- read.delim("sample6_peaks.narrowPeak", header=FALSE)
 all_peak <- rbind(sample1_peaks, sample2_peaks, sample3_peaks, sample4_peaks, sample5_peaks, sample6_peaks)
 peaks_merged_gtf <- merge_peak(all_peak)
}
}
