% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bed.R
\name{get_bed}
\alias{get_bed}
\title{Extract 'chr', 'start', 'end' columns of peak file}
\usage{
get_bed(peak)
}
\arguments{
\item{peak}{peak file which should contain 'chr', 'start', 'end' columns}
}
\value{
return bed format data frame
}
\description{
Extract 'chr', 'start', 'end' columns of peak file
}
