% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_footprints.R
\name{get_merged_fasta}
\alias{get_merged_fasta}
\title{merge footprints and return corresponding sequence-}
\usage{
get_merged_fasta(footprints, fastadir, distance = 4)
}
\arguments{
\item{footprints}{footprints file, generated by dnase2tf or Hint or other related software, name of first column should be 'chr', name of second column should be 'start' and name of thrid column should be 'end'.}

\item{fastadir}{character, indiating the path of reference genome}

\item{distance}{numeric, indicating the cutoff for merging footprints}
}
\value{
return sequence of footprints in fasta format
}
\description{
This function first merges footprints whose distances is less than 4, and then obtains the sequence of each footprint based on the reference genome, and  output the sequence in FASTA format
}
\examples{
fastadir='Genome/GRCm38Chr.fasta'
#merged_fasta <- get_merged_fasta(fdr005,fastadir)
}
