% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scATAC.R
\name{Get_p2g_fun}
\alias{Get_p2g_fun}
\title{Filter peak-to-gene link}
\usage{
Get_p2g_fun(
  ArchR_obj,
  corCutOff = 0.2,
  FDRCutOff = 1e-06,
  varCutOffATAC = 0.7,
  varCutOffRNA = 0.3
)
}
\arguments{
\item{ArchR_obj}{ArchR object with Peak2GeneLinks information in peakset slot}

\item{corCutOff}{numeric, correlation between each peak and gene to refine p2g}

\item{FDRCutOff}{numeric, FDR of peak to gene link to refine p2g}

\item{varCutOffATAC}{numeric, VarQATAC to refine p2g}

\item{varCutOffRNA}{numeric, VarQRNA to refine p2g}
}
\value{

}
\description{
Extract peak to gene links data.frame from ArchR object, then identify
significant links based on corCutOff, FDRCutOff, varCutOffATAC and varCutOffRNA.
}
