% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct_network.R
\name{get_partial_regulations}
\alias{get_partial_regulations}
\title{Reconstructing integrated regulatory networks}
\usage{
get_partial_regulations(TFs_list)
}
\arguments{
\item{TFs_list}{list, indicating enriched TFs list which generated by get_Enriched_TFs()}
}
\value{
return regulatory network for enriched TFs of each module
}
\description{
Reconstructing integrated regulatory networks
}
\examples{
tf_network <- get_partial_regulations(TFs_list)
}
