% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_peaks.R
\name{get_peaks_genes}
\alias{get_peaks_genes}
\title{Get candidate genes/TFs-related peaks}
\usage{
get_peaks_genes(list1, expression_profile)
}
\arguments{
\item{list1}{list, generated by get_related_genes(), the first element should be bed data.frame(contain 3 columns), the second element should contain 9 columns}

\item{expression_profile}{dataframe, indicating expression profile of bulk RNA-seq or single-cell RNA-seq.}
}
\value{
return a list contain two dataframe, first one is candidate genes/TFs-related peaks dataframe, second one is bed dataframe
}
\description{
Get candidate genes/TFs-related peaks
}
\examples{
list1 <- get_peaks_genes(list1,MmscRNA_PHx_Exp_NewF)
}
