% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Footprints_FOS.R
\name{get_potential_regulation}
\alias{get_potential_regulation}
\title{Inner function}
\usage{
get_potential_regulation(FOSF)
}
\arguments{
\item{FOSF}{filtered FOS}
}
\value{
return regulatory relationships
}
\description{
Inner function
}
