% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct_network.R
\name{get_regulation_of_TFs_to_modules}
\alias{get_regulation_of_TFs_to_modules}
\title{Get regulation of TFs to modules}
\usage{
get_regulation_of_TFs_to_modules(TFs_list, Thr = 2)
}
\arguments{
\item{TFs_list}{list, indicating enriched TFs list which generated by get_Enriched_TFs()}

\item{Thr}{numeric to filter enriched transcription factors, default is 2}
}
\value{
add TFs and TFs to module regulation relationship to TFs_list
}
\description{
Get regulation of TFs to modules
}
\examples{
TFs_list = get_regulation_of_TFs_to_modules(TFs_list, Thr = 2)

}
