% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{get_related_genes}
\alias{get_related_genes}
\title{Annotate peaks}
\usage{
get_related_genes(footprints, motif, Species, txdb, tssRegion = c(-3000, 3000))
}
\arguments{
\item{footprints}{footprints file, generated by overlap_footprints_peaks() or intersect function of bedtools}

\item{motif}{motif file, you can choose our bulit-in motif database of 'mus musculus', 'homo sapiens', 'zebrafish' and 'chicken' by 'motif = Tranfac201803_Mm_MotifTFsF', 'motif = Tranfac201803_Hs_MotifTFsF', 'motif = Tranfac201803_Zf_MotifTFsF', 'motif = Tranfac201803_Ch_MotifTFsF' respectively, or you can upload your own motif data base, but the formata use be the same as our built-in motif database.}

\item{Species}{character, indicating the species of data which is used to choose annodb when annotate peak.}

\item{txdb}{TxDb object contained transcript-related features of a particular genome. Bioconductor provides several package that containing TxDb object of model organisms with multiple commonly used genome version, for instance TxDb.Hsapiens.UCSC.hg38.knownGene, TxDb.Hsapiens.UCSC.hg19.knownGene for human genome hg38 and hg19, TxDb.Mmusculus.UCSC.mm10.knownGene and TxDb.Mmusculus.UCSC.mm9.knownGene for mouse genome mm10 and mm9, etc.}

\item{tssRegion}{Region Range of TSS}
}
\value{
return a list, first element is bed format datafrmae, second element is annotated footprints dataframe
}
\description{
This function first merge and extend footprint regions, and then integrate R package ChIPseeker to get footprint-related genes
}
\examples{
#txdb <- TxDb.Mmusculus.UCSC.mm10.knownGene::TxDb.Mmusculus.UCSC.mm10.knownGene
load(system.file("extdata", "combined.rda", package = "IReNA"))
load(system.file("extdata", "test_peak.rda", package = "IReNA"))
peak_bed <- get_bed(test_peak)
overlapped <- overlap_footprints_peaks(combined, peak_bed)
#list1 <- get_related_genes(overlapped,txdb = txdb,motif=Tranfac201803_Mm_MotifTFsF,Species = 'Mm')
}
