% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATAC-seq.R
\name{get_related_peaks}
\alias{get_related_peaks}
\title{Get candidate genes/TFs-related peaks}
\usage{
get_related_peaks(list1, expression_profile)
}
\arguments{
\item{list1}{list, generated by get_related_genes(), the first element should
be bed data.frame(contain 3 columns), the second element should contain 9 columns}

\item{expression_profile}{dataframe, indicating expression profile of bulk
RNA-seq or single-cell RNA-seq.}
}
\value{
return a list contain two dataframe, first one is candidate genes/TFs-related
peaks dataframe, second one is bed dataframe
}
\description{
refine peaks through genes in expression profile of scRNA-seq data
}
\examples{
load(system.file("extdata", "list1.rda", package = "IReNA"))
load(system.file("extdata", "test_clustering.rda", package = "IReNA"))
Kmeans_clustering_ENS <- add_ENSID(test_clustering, Spec1='Hs')
list1 <- get_related_peaks(list1,Kmeans_clustering_ENS)
}
