% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_fimo.R
\name{get_tss_region}
\alias{get_tss_region}
\title{Get tss region sequence of target genes in regulaotry relationships}
\usage{
get_tss_region(gtf, gene.use, upstream_length = 1000, downstream_length = 500)
}
\arguments{
\item{gtf}{Gene transfer format, you can download it from http://www.ensembl.org/info/data/ftp/index.html}

\item{gene.use}{character, indicating target genes}

\item{upstream_length}{numeric, indicating upstream region length of TSS}

\item{downstream_length}{numeric, indicating downstream region length of TSS}
}
\value{
data.frame, contains four columns are target gene, chr, start and end.
}
\description{
Get tss region sequence of target genes in regulaotry relationships
}
