% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_footprints.R
\name{getfasta}
\alias{getfasta}
\title{Inner function to get fasta from each footprints}
\usage{
getfasta(merged_footprints, fastadir)
}
\arguments{
\item{merged_footprints}{merged footprints, generated by merge_footprints()}

\item{fastadir}{reference genome}
}
\value{

}
\description{
Inner function to get fasta from each footprints
}
