% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{HsscRNA_genes}
\alias{HsscRNA_genes}
\title{homo sapiens Gene ID correspondence}
\format{
contain ENSEMBEL ID, Symbol ID and NCBI ID
}
\usage{
HsscRNA_genes
}
\description{
homo sapiens Gene ID correspondence
}
\keyword{datasets}
