% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_pipeline.R
\name{irena_step1}
\alias{irena_step1}
\title{irena_step1}
\usage{
irena_step1(obj, metacell_num = 50, module_number = 4)
}
\arguments{
\item{obj}{seurat object, where meta.data should contain Pseudotime column}

\item{metacell_num}{numeric, indicating excepted metacell number}

\item{module_number}{numeric,indicating module number}
}
\value{
return a data.frame, and the first column is K-means group
}
\description{
irena_step1
}
