% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_pipeline.R
\name{irena_step2}
\alias{irena_step2}
\title{Title}
\usage{
irena_step2(
  exp,
  grn_method = "pearson correlation",
  start_column = 2,
  nCores = 5,
  candidate_tfs
)
}
\arguments{
\item{candidate_tfs}{}
}
\description{
Title
}
