% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{list1}
\alias{list1}
\title{data for demonstrating}
\format{
a list contain 2 datafrmame, the first one is footprints and the second one is bed format footprints
}
\usage{
list1
}
\description{
data for demonstrating
}
\keyword{datasets}
