% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA1.R
\name{load_counts}
\alias{load_counts}
\title{load counts}
\usage{
load_counts(datapath, datatype = 0)
}
\arguments{
\item{datapath}{character, indicating data path of counts, if it is 10X data, please input the path of folder which containing matrix.mtx.gz, features.tsv.gz and barcodes.tsv.gz}

\item{datatype}{10X:datatype = 0, counts:datatype = 1, sparse matrix:datatype = 2}
}
\value{
return a seurat object
}
\description{
Load counts from datapath and return a seurat object
}
\examples{
\dontrun{load_counts('D:/scRNA/10X',datatype=1)}
}
