% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_footprints.R
\name{merge_footprints}
\alias{merge_footprints}
\title{Inner function to merge footprints whose distance is less than 4}
\usage{
merge_footprints(footprints, ditance = 4, revise = TRUE)
}
\arguments{
\item{footprints}{footprints file less than distance}

\item{ditance}{default is 4}

\item{revise}{whether revise the footprints}
}
\description{
Inner function to merge footprints whose distance is less than 4
}
