% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct_network.R
\name{merge_Module_Regulations}
\alias{merge_Module_Regulations}
\title{Constructing intramodular regulatory networks}
\usage{
merge_Module_Regulations(TFs_list, Kmeans_result, ModuleThr1 = 0.05)
}
\arguments{
\item{TFs_list}{list, indicating enriched TFs list which generated by get_Enriched_TFs()}

\item{Kmeans_result}{Kmeans result data.frame, row names should be ENSEMBEL ID}

\item{ModuleThr1}{FDR to filter significant regulations default = 0.05}
}
\value{
return intramodular network
}
\description{
Constructing intramodular regulatory networks
}
\examples{
intramodular_network <- merge_Module_Regulations(TFs_list, Kmeans_clustering, ModuleThr1 = 0.05)

}
