% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Footprints_FOS.R
\name{Merge_same_pairs}
\alias{Merge_same_pairs}
\title{Inner function}
\usage{
Merge_same_pairs(FOSF_Reg)
}
\arguments{
\item{FOSF_Reg}{FOS with potential regulatory}
}
\value{
merged pairs
}
\description{
Inner function
}
