% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{merge_sort_count}
\alias{merge_sort_count}
\title{Merge and sort counts}
\usage{
merge_sort_count(count_all, gtf)
}
\arguments{
\item{count_all}{data.frame, indicating counts of peaks, generated by htseq-count.
If you calculate counts for each sample separately, you need to combine these counts firstly through cbind().}

\item{gtf}{data.frame, indicating gtf file of all peaks, generated by
\link{generate_gtf} function}
}
\value{
return sorted counts
}
\description{
Sort the peak files by end of peak, start of peak, and chr of peak
}
\examples{
\dontrun{
SSC_patient1_Counts <- read.delim("SSC_patient1_Counts.txt", header=FALSE)
SSC_patient2_Counts <- read.delim("SSC_patient2_Counts.txt", header=FALSE)
esc_Counts_Counts <- read.delim("esc_Counts.txt", header=FALSE)
peaks_gtf <- read.delim("peak_merged.gtf", header=FALSE)
count_all <- cbind(SSC_patient1_Counts, SSC_patient2_Counts[,2],
             esc_Counts_Counts[,2])
merged_count <- merge_sort_count(count_all, peaks_gtf)
}

}
