% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Footprints_FOS.R
\name{Merge_TFs_genes}
\alias{Merge_TFs_genes}
\title{Inner function}
\usage{
Merge_TFs_genes(FOSF_RegM)
}
\arguments{
\item{FOSF_RegM}{merged FOS with potential regulatory}
}
\description{
Inner function
}
