% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{MmscRNA_genes}
\alias{MmscRNA_genes}
\title{Mus musculus Gene ID correspondence}
\format{
contain ENSEMBEL ID, Symbol ID and NCBI ID
}
\usage{
MmscRNA_genes
}
\description{
Mus musculus Gene ID correspondence
}
\keyword{datasets}
