% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{MmscRNA_PHx_Exp_NewF}
\alias{MmscRNA_PHx_Exp_NewF}
\title{data for test}
\format{
expression profile, first column is ENSEMBLE ID, second column is Symbol
}
\usage{
MmscRNA_PHx_Exp_NewF
}
\description{
data for test
}
\keyword{datasets}
