% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif_check.R
\name{motifs_check}
\alias{motifs_check}
\title{Title}
\usage{
motifs_check(motif, Dir)
}
\arguments{
\item{motif}{motif file}

\item{Dir}{motif directory contain motif files generated by Fimo}
}
\value{
print motif which do not include in the file
}
\description{
Title
}
