% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct_network.R
\name{network_analysis}
\alias{network_analysis}
\title{Function to make regulatory network analysis}
\usage{
network_analysis(
  regulatory_relationships,
  Kmeans_result,
  TFFDR1 = 10,
  TFFDR2 = 10,
  ModuleFDR = 0.05
)
}
\arguments{
\item{regulatory_relationships}{Refined or unrefined regulatory relationships.
Unrefined regulatory relationships are generated by \code{\link{get_cor}},
and you can use \code{\link{filter_regulation}} or consequence of
\code{\link{Footprints_FOS}} to refine it}

\item{Kmeans_result}{Kmeans result data.frame, row names should be ENSEMBEL ID,
and the first column should be gene Symbol ID, the second column should be KmeansGroup}

\item{TFFDR1}{numeric, indicating the cutoff (FDR) to identify enriched transcription factors}

\item{TFFDR2}{numeric, indicating the cutoff (FDR) to identify enriched
transcription factors that regulate each module to make regulatory network
for enriched transcription factors of each module}

\item{ModuleFDR}{numeric, indicating the cutoff (FDR) to identify enriched
modular regulatory relationships to make intramodular regulatory network}
}
\value{
This function return a list which contain 10 element

Cor_TFs: data.frame of expressed TFs in the gene networks

Cor_EnTFs: data.frame of TFs which significantly regulate gene modules (or enriched TFs)

FOSF_RegMTF_Cor_EnTFs: regulatory pairs in which the source gene is enriched TF

FOSF_RegMTF_Cor_EnTFsTarg: regulatory pairs in which both source gene and target gene are enriched TFs

FOSF_RegMTF_Cor_EnTFsTargM: regulatory pairs only including regulations within each module but not those between modules

TF_list: enriched TFs which significantly regulate gene modules

TF_module_regulation: details of enriched TFs which significantly regulate gene modules

TF_network: regulatory network for enriched transcription factors of each module

intramodular_network: intramodular regulatory network
}
\description{
This function calculates the significant transcription factors
and significant regulatory relationships in each module through hypergeometric
tests to construct regulatory networks
}
\examples{
load(system.file("extdata", "test_clustering.rda", package = "IReNA"))
Kmeans_cluster_Ens <- add_ENSID(test_clustering,Spec1='Hs')
motif1 <- Tranfac201803_Hs_MotifTFsF
regulatory_relationships <- get_cor(Kmeans_cluster_Ens,motif1,0.7)
TFs_list <- network_analysis(regulatory_relationships,Kmeans_cluster_Ens)
}
