% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct_network.R
\name{output_grn}
\alias{output_grn}
\title{Generate grn format regulatory relationships}
\usage{
output_grn(regulatory_relationships)
}
\arguments{
\item{regulatory_relationships}{regulatory relationships , columns should
contain 'TF', 'TFSymbol', 'TFGroup', 'Target', 'TargetSymbol', 'TargetGroup'
and 'Correlation'}
}
\value{
grn format regulatory relationships
}
\description{
Generate grn format regulatory relationships
}
\examples{
load(system.file("extdata", "test_clustering.rda", package = "IReNA"))
Kmeans_cluster_Ens <- add_ENSID(test_clustering,Spec1='Hs')
motif1 <- Tranfac201803_Hs_MotifTFsF
regulatory_relationships <- get_cor(Kmeans_cluster_Ens,motif1,0.9)
grn <- output_grn(regulatory_relationships)
}
