% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_footprints.R
\name{overlap_footprints_peaks}
\alias{overlap_footprints_peaks}
\title{Overlap differential peaks and motif footprints}
\usage{
overlap_footprints_peaks(footprints, peak_bed)
}
\arguments{
\item{footprints}{footprints file, generated by combine_footprints()}

\item{peak_bed}{bed format of peak, first column should be 'chr', second column should be 'start', third column should be 'end'}
}
\value{
return data.frame contain footprints that overlap peaks
}
\description{
Overlap differential peaks and motif footprints
}
\examples{
peak_bed <- get_bed(peak)
overlapped <- overlap_footprints_peaks(combined, peak_bed)
}
