% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_visualization.R
\name{plot_intramodular_network}
\alias{plot_intramodular_network}
\title{Function to visualize intramodular regulatory network}
\usage{
plot_intramodular_network(
  TFs_list,
  enrichment = NULL,
  layout = "circle",
  group.cols = NULL,
  title.name = NULL,
  vertex.size = 25,
  vertex.label.color = "black",
  edge.label.color = "black",
  legend = TRUE,
  vertex.label.cex = 1.1,
  vertex.label.family = "ArialMT",
  frame.color = "white",
  arrow.size = 0.5,
  arrow.width = 0.8,
  edge.width = 2.5,
  edge.curved = 0,
  edge.color = c("#FDD1B0", "#B3B3B3"),
  vertex.label.degree = 0.35,
  vertex.label.dist = -6
)
}
\arguments{
\item{TFs_list}{TFs_list generated by \code{\link{network_analysis}}}

\item{enrichment}{Enrichment analysis results of genes in each module,
generated by \code{\link{enrich_module}}. If it is not NULL, when plotting the
Intramodular Network, the functions enriched by each module will be plotted
in the diagram}

\item{layout}{the layout to display the network, options: 'grid','sphere',
'circle','random'}

\item{group.cols}{colors for group in network}

\item{title.name}{the name of the title}

\item{vertex.size}{vertex size}

\item{vertex.label.color}{ertex label color}

\item{edge.label.color}{edge label color}

\item{legend}{logic, indicating whether to show the legend}

\item{vertex.label.cex}{The label size of vertex}

\item{vertex.label.family}{vertex label family}

\item{frame.color}{frame.color}

\item{arrow.size}{arrow size}

\item{arrow.width}{arrow width}

\item{edge.width}{edge width}

\item{edge.curved}{edge curvature}

\item{edge.color}{edge color. You need to input two colors, first one indicate
'Positive' regulation, second one indicate 'Negative' regulation.}

\item{vertex.label.degree}{The position of the label in relation to the vertex
(This parameter is valid only when the enrichment does not equal null)}

\item{vertex.label.dist}{Distance between the label and the vertex
(This parameter is valid only when the enrichment does not equal null)}
}
\value{
figure
}
\description{
This function integrate R package igraph to visualize consequence
of regulatory network analysis
}
\examples{
#'
load(system.file("extdata", "test_clustering.rda", package = "IReNA"))
Kmeans_clustering <- add_ENSID(test_clustering, Spec1 = "Hs")
cor0.6 <- get_cor(Kmeans_clustering, Tranfac201803_Hs_MotifTFsF, 0.6, start_column=3)
TFs_list <- network_analysis(cor0.6,Kmeans_clustering)
#enrichment <- enrich_module(Kmeans_cluster_Ens, org.Hs.eg.db, 'KEGG')
#plot_intramodular_network(TFs_list,enrichment,layout = 'random')
}
