% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth.R
\name{plot_kmeans_pheatmap}
\alias{plot_kmeans_pheatmap}
\title{plot kmeans pheatmap}
\usage{
plot_kmeans_pheatmap(
  Kmeans_result,
  start_column = 3,
  Gene1 = NULL,
  NumRowBlank1 = 20,
  show_colnames = FALSE,
  clustering_distance_rows = "correlation",
  ModuleColor1 = NULL,
  ModuleScale1 = 20,
  cluster_cols = FALSE,
  fontsize = 10,
  legend1 = TRUE,
  border_color = "white",
  ByPanel = FALSE,
  Color1 = NULL,
  Show.Module = TRUE,
  Range1 = c(-3, 3),
  ColumnGroup1 = NULL
)
}
\arguments{
\item{Kmeans_result}{expression profile which should contain column "KmeansGroup"}

\item{start_column}{numeric, indicating the start column of expression value,
defalut is 3}

\item{Gene1}{custom labels for rows that are used instead of rownames}

\item{NumRowBlank1}{the blank between each group}

\item{show_colnames}{logic, indicating whether show column names, default is FALSE}

\item{clustering_distance_rows}{distance measure used in clustering rows.
Possible values are "correlation" for Pearson correlation and all the
distances supported by dist, such as "euclidean", etc. If the value
is none of the above it is assumed that a distance matrix is provided}

\item{ModuleColor1}{each color for each module}

\item{ModuleScale1}{change the relative proportion of module bar and expression heatmap}

\item{cluster_cols}{boolean values determining if columns should be clustered or hclust object.}

\item{fontsize}{font size}

\item{legend1}{logic, indicating whether show the legend, default is TRUE}

\item{border_color}{color of cell borders on heatmap, use NA if no border
should be drawn.}

\item{ByPanel}{logic, if TRUE, reorder columns of heatmap according to the
parameter ColumnGroup1, default is FALSE}

\item{Color1}{colors used in heat map}

\item{Show.Module}{logic, indicating whether show the module, defalut is TURE}

\item{Range1}{numeric, indicating the range of expression values which will
affect the range of heat map, defalut is c(-3, 3)}

\item{ColumnGroup1}{vector, if it is not NULL, this function wil sepate columns
according to varible ColumnGroup1, default is NULL}
}
\value{
Kmeans clustering figure
}
\description{
Display the result of clustering based on pheatmap, the input
data.frame should contain KmeansGroup column.
}
\examples{
col1 <- c('#67C1E3','#EF9951','#00BFC4','#AEC7E8','#C067A9','#E56145','#2F4F4F')
load(system.file("extdata", "test_clustering.rda", package = "IReNA"))
plot_kmeans_pheatmap(test_clustering, ModuleColor1 = col1,Range1=c(-3,3),NumRowBlank1=1,ModuleScale1 = 20)

}
