% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_visualization.R
\name{plot_network}
\alias{plot_network}
\title{Function to visualize regulatory network}
\usage{
plot_network(
  TFs_list,
  enrichment = NULL,
  layout,
  group.cols = NULL,
  title.name = NULL,
  vertex.size = 10,
  type = "TF",
  vertex.label.color = "black",
  edge.label.color = "black",
  legend = TRUE,
  vertex.label.cex = 0.6,
  vertex.label.family = "ArialMT",
  frame.color = "white",
  arrow.size = 0.2,
  arrow.width = 0.5,
  edge.with = 1.8,
  edge.curved = 0,
  edge.color = "grey",
  vertex.label.degree = 0.35,
  vertex.label.dist = -6
)
}
\arguments{
\item{TFs_list}{TFs_list generated by \code{\link{network_analysis}}}

\item{enrichment}{Enrichment analysis results of genes in each module, generated by \code{\link{enrichment_analysis}}. If it is not NULL, when plotting the Intramodular Network, the functions enriched by each module will be plotted in the diagram}

\item{layout}{the layout to display the network, options: 'grid','sphere','circle','random'}

\item{group.cols}{colors for group in network}

\item{title.name}{the name of the title}

\item{vertex.size}{vertex size}

\item{type}{network type, 'TF' indicate TFs network, 'module' indicate intramodular network}

\item{vertex.label.color}{ertex label color}

\item{edge.label.color}{edge label color}

\item{legend}{logic, indicating whether to show the legend}

\item{vertex.label.cex}{The label size of vertex}

\item{vertex.label.family}{vertex label family}

\item{frame.color}{frame.color}

\item{arrow.size}{arrow size}

\item{arrow.width}{arrow width}

\item{edge.with}{edge width}

\item{edge.curved}{edge curvature}

\item{edge.color}{edge color}

\item{vertex.label.degree}{The position of the label in relation to the vertex (This parameter is valid only when the enrichment does not equal null)}

\item{vertex.label.dist}{Distance between the label and the vertex (This parameter is valid only when the enrichment does not equal null)}
}
\value{
figure
}
\description{
This function integrate R package igraph to visualize consequence of regulatory network analysis
}
\examples{
#' Kmeans_clustering <- add_ENSID(Kmeans_clustering, Spec1 = "Hs")
cor0.6 <- get_cor(Kmeans_clustering, "Hs", 0.9, start_column=3)
TFs_list <- network_analysis(regulatory_relationships,Kmeans_cluster_Ens)
plot_network(TFs_list,layout = 'circle',type = 'TF')
plot_network(TFs_list,layout = 'random',type = 'module',vertex.size = 25,vertex.label.cex = 1.1,edge.with = 2,arrow.size = 0.5)
#enrichment <- enrich_module(Kmeans_cluster_Ens, org.Hs.eg.db, 'KEGG')
#plot_network(TFs_list,enrichment,layout = 'random',type = 'module',vertex.size = 25,vertex.label.cex = 1.1,edge.with = 2,arrow.size = 0.5,vertex.label.family = 'Times')
}
