% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth.R
\name{smoothByBin}
\alias{smoothByBin}
\title{Internal function}
\usage{
smoothByBin(
  Exp1,
  Pseudotime1,
  PseudotimeRange1 = NULL,
  SmoothLength1 = 100,
  ByBin1 = c("Equal.Pseudotime", "Equal.Cells")
)
}
\arguments{
\item{Exp1}{the raw data, such as seurat_object@assays$RNA@data}

\item{Pseudotime1}{meta.data of seurat object}

\item{PseudotimeRange1}{give the set pseudotime range}

\item{SmoothLength1}{the bins which divide the pseudotime}

\item{ByBin1}{function used, Equal.Pseudotime and Equal.Cells}
}
\value{
return a expression profile
}
\description{
Internal function
}
