% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct_network.R
\name{sort_TFs_degree}
\alias{sort_TFs_degree}
\title{Sort Transcription factor based on degree.}
\usage{
sort_TFs_degree(regulatory_relationships)
}
\arguments{
\item{regulatory_relationships}{}
}
\value{
Matrix contains sorted transcription factors
}
\description{
Sort Transcription factor based on degree.
}
\examples{
load(system.file("extdata", "test_clustering.rda", package = "IReNA"))
correlation <- get_cor(test_clustering, Tranfac201803_Hs_MotifTFsF, 0.7, start_column=3)
TFs_degree <- sort_TFs_degree(correlation)
}
