% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_dataframe.R
\name{split_dataframe}
\alias{split_dataframe}
\title{inner function to split dataframe}
\usage{
split_dataframe(dataframe1, sep = "\\t")
}
\arguments{
\item{dataframe1}{dataframe need to be splited}

\item{sep}{split symbol}
}
\description{
inner function to split dataframe
}
