% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{Tranfac201803_Mm_MotifTFsF}
\alias{Tranfac201803_Mm_MotifTFsF}
\title{Tranfac201803 database of chicken}
\format{
contain motifs and its corresponding genes
}
\usage{
Tranfac201803_Mm_MotifTFsF
}
\description{
Tranfac201803 database of chicken
}
\keyword{datasets}
