% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{Tranfac201803_MotifPWM}
\alias{Tranfac201803_MotifPWM}
\title{Tranfac201803 database of motif Position weight matrix}
\format{
An object of class \code{data.frame} with 120486 rows and 1 columns.
}
\usage{
Tranfac201803_MotifPWM
}
\description{
Tranfac201803 database of motif Position weight matrix
}
\keyword{datasets}
