% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Footprints_FOS.R
\name{Trans_WigToMultirows2}
\alias{Trans_WigToMultirows2}
\title{Inner function}
\usage{
Trans_WigToMultirows2(b)
}
\arguments{
\item{b}{wig}
}
\value{
return transfered wig file
}
\description{
Inner function
}
