% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{ZfscRNA_genes}
\alias{ZfscRNA_genes}
\title{Zebrafish Gene ID correspondence}
\format{
contain ENSEMBEL ID, Symbol ID and NCBI ID
}
\usage{
ZfscRNA_genes
}
\description{
Zebrafish Gene ID correspondence
}
\keyword{datasets}
