% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA1.R
\name{add_pseudotime_DEG_filter}
\alias{add_pseudotime_DEG_filter}
\title{add pseudtoime and identify DEGs}
\usage{
add_pseudotime_DEG_filter(
  seurat_object,
  monocle_object,
  DEG = TRUE,
  qvalue = 0.05,
  nce = 0.1,
  ed = 0.1,
  normlize1 = TRUE
)
}
\arguments{
\item{seurat_object}{seurat object, which has same cells and genes as monocle_object}

\item{monocle_object}{monocle object, which include pesudotime and has same
cells and genes as seurat_object}

\item{DEG}{logic, indicating whether filter differential expressed genes for
seurat object}

\item{qvalue}{numeric, indicating q-value to indentify differentially
expressed genes}

\item{nce}{numeric, indicating num cells expressed to indentify differentially
expressed genes}

\item{ed}{numeric, indicating expression difference to indentify differentially
expressed genes}

\item{normlize1}{logic, indicating whether normalize the data in seurat object}
}
\value{
return a seurat object
}
\description{
Add pseudotime to seurat object, and filter differential
expressed genes according to pseudotime
}
\examples{
load(system.file("extdata", "test_seurat.rda", package = "IReNA"))
monocle_object = get_pseudotime(test_seurat)
add_pseudotime_DEG_filter(seurat_object = test_seurat,monocle_object = monocle_object, DEG = FALSE, normlize1 = FALSE)
#add_pseudotime_DEG_filter(seurat_object = test_seurat,monocle_object = monocle_object, DEG = TRUE, qvalue = 0.001, nce = 0.1, ed = 0.1)
}
