% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATAC-seq.R
\name{combine_footprints}
\alias{combine_footprints}
\title{Combine all footprints of motifs}
\usage{
combine_footprints(Dir)
}
\arguments{
\item{Dir}{character, indicating the path of all footprints related motif which generated by fimo. Folder in this path cannot contain any other irrelevant files}
}
\value{
return combined footprints
}
\description{
combine all footprints related motifs which generated by fimo. Before run this function you should check whether the 'Dir' only contain consequence of fimo
}
\examples{
#combied<-combine_footprints(Dir2)
}
