% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth.R
\name{fileter_expression_profile}
\alias{fileter_expression_profile}
\title{fileter expression profile}
\usage{
fileter_expression_profile(expression_profile, filterfc = TRUE, FC = 0.1)
}
\arguments{
\item{expression_profile}{expression profile, generated by
\code{\link{get_SmoothByBin_PseudotimeExp}}}

\item{filterfc}{logic, indicating whether you want to filter genes
based on logFC}

\item{FC}{numeric, indicating logFC threshold to filter genes}
}
\value{
return filtered expression profile
}
\description{
this function is used to filter the bins in the expression
profile which value is zero
}
\examples{
load(system.file("extdata", "test_clustering.rda", package = "IReNA"))
expression_profile = test_clustering[,-1]
fileter_expression_profile(expression_profile)
}
