% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_motifs.R
\name{find_motifs}
\alias{find_motifs}
\title{Find motifs in the footprints through fimo software.}
\usage{
find_motifs(motif, step = 20, Dir, sequence_dir)
}
\arguments{
\item{motif}{motif file, you can choose our bulit-in motif database of 'mus musculus', 'homo sapiens', 'zebrafish' and 'chicken' by 'motif = Tranfac201803_Mm_MotifTFsF', 'motif = Tranfac201803_Hs_MotifTFsF', 'motif = Tranfac201803_Zf_MotifTFsF', 'motif = Tranfac201803_Ch_MotifTFsF' respectively, or you can upload your own motif data base, but the formata use be the same as our built-in motif database.}

\item{step}{numeric, indicating the numbers of motif in each group. Because there are so many motifs need to be calculated, so we divided every 20 motifs into groups and then calculated each group using FIMO at the same time by nohup}

\item{Dir}{characater, indicating the directory of output files}

\item{sequence_dir}{sequence file directory}
}
\value{
return scripts to run Fimo in command line
}
\description{
Find motifs in the footprints through fimo software.
}
