% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Footprints_FOS.R
\name{Footprints_FOS}
\alias{Footprints_FOS}
\title{Calculate FOS of footprints}
\usage{
Footprints_FOS(Wig_list, Candid, FOS_threshold = 1, trans_wig = FALSE)
}
\arguments{
\item{Wig_list}{list object, where each element is the \link{wig_track} result of each sample.}

\item{Candid}{If you follow our our pipline, this parameter should be the second element of the list that generated by get_peaks_genes()}

\item{FOS_threshold}{numeric, indicating the threshold of footprint occupancy score to identify significant footprints.}

\item{trans_wig}{logcial, if you use dnase_wig_tracks_both2.py to get wig_list, please set trans_wig parameter as TRUE.}
}
\value{
return related transcription factors of footprints with FOS above FOS_threshold
}
\description{
Calculate the FOS(footprints occupancy score) to identify enriched transcriptions factor, and use these transcription factors to identify significant regulatory relationships
}
\examples{
Kmeans_clustering_ENS <- add_ENSID(clustering, Spec1='Hs')
regulatory_relationships <- Footprints_FOS(wig_list, Candid, Kmeans_clustering_ENS)
}
