% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATAC-seq.R
\name{get_bed}
\alias{get_bed}
\title{get bed}
\usage{
get_bed(peak)
}
\arguments{
\item{peak}{peak file which should contain 'chr', 'start', 'end' columns}
}
\value{
return bed format data frame
}
\description{
Extract 'chr', 'start', 'end' columns of peak file
}
\examples{
load(system.file("extdata", "test_peak.rda", package = "IReNA"))
peak_bed <- get_bed(test_peak)
}
