% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct_network.R
\name{get_Enriched_TFs}
\alias{get_Enriched_TFs}
\title{Get enriched transcription factors}
\usage{
get_Enriched_TFs(GeneCor1, Kmeans_result, TFFdrThr1 = 2)
}
\arguments{
\item{GeneCor1}{correlation data.frame, yield by get_cor}

\item{Kmeans_result}{Kmeans result data.frame, row names should be ENSEMBEL ID}

\item{TFFdrThr1}{numeric to filter enriched transcription factors, defalut is 2}
}
\value{
return enriched TFs list which contain 5 dataframe,
"Cor_TFs" is list of expressed TFs in the gene networks,
"Cor_EnTFs" is list of TFs which significantly regulate gene modules (or enriched TFs),
"FOSF_RegMTF_Cor_EnTFs" is regulatory pairs in which the source gene is enriched TF,
"FOSF_RegMTF_Cor_EnTFsTarg" is regulatory pairs in which both source gene and target gene are enriched TFs,
"FOSF_RegMTF_Cor_EnTFsTargM" is regulatory pairs only including regulations within each module but not those between modules
}
\description{
Calculate enriched transcription factors in each module through hypergeometric test
}
\examples{
Kmeans_clustering <- add_ENSID(Kmeans_clustering, Spec1 = "Hs")
cor0.6 <- get_cor(Kmeans_clustering, "Hs", 0.9, start_column=3)
TFs_list <- get_Enriched_TFs(cor0.6, Kmeans_clustering, TFFdrThr1 = 2, correlation_filter = 0)

}
