% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_footprints.R
\name{get_merged_fasta}
\alias{get_merged_fasta}
\title{This function first merges footprints whose distances is less than 4, and then obtains the sequence of each footprint based on the reference genome, and  output the sequence in FASTA format}
\usage{
get_merged_fasta(footprints, distance = 4, fastadir)
}
\arguments{
\item{footprints}{footprints file, generated by dnase2tf() function}

\item{distance}{numeric, indicating the cutoff for merging footprints}

\item{fastadir}{character, indiating the path of reference genome}
}
\value{
return sequence of footprints in fasta format
}
\description{
This function first merges footprints whose distances is less than 4, and then obtains the sequence of each footprint based on the reference genome, and  output the sequence in FASTA format
}
