% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRNA1.R
\name{get_pseudotime}
\alias{get_pseudotime}
\title{Calculate pseudotime of cells}
\usage{
get_pseudotime(seurat_object, reverse = TRUE, show_trajecotry = TRUE)
}
\arguments{
\item{seurat_object}{seurat object}

\item{reverse}{TRUE or FALSE, whether to reverse the pseudotime, default is TURE}

\item{show_trajecotry}{TRUE or FALSE, whether show the trajectory of pseudotime, default is TURE}
}
\value{
return monocle object which contain pseudotime
}
\description{
Use monocle to calculate the pseudotime and return a monocle object
}
\examples{
load(system.file("extdata", "test_seurat.rda", package = "IReNA"))
get_pseudotime(test_seurat)
}
