% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth.R
\name{get_SmoothByBin_PseudotimeExp}
\alias{get_SmoothByBin_PseudotimeExp}
\title{Smooth cells into bins based on pseudotime or State}
\usage{
get_SmoothByBin_PseudotimeExp(
  seurat_object,
  FC = TRUE,
  Bin = 50,
  method = "Pseudotime",
  FcType = "Q95"
)
}
\arguments{
\item{seurat_object}{seurat object, where meta.data should contain
Pseudotime}

\item{FC}{logic, indicating whether to add FoldChangeQ95 to the first column,
default is TRUE}

\item{Bin}{numeric, indicating the numbers of bins which divide the pseudotime,
default is 50}

\item{method}{'Pseudotime' or 'State'. Pseudotime method firstly orders all
cells according to pseudotime, and then smooth these cells into ‘Bin' number
of bins. State method firstly orders all cells according to pseudotime, and
then smooth cells in each state into ('Bin'/number of state) number of bins.}

\item{FcType}{'Q95' or 'Q90', FoldChange threshold, default is 'Q95'}
}
\value{
return a expression profile
}
\description{
Divide cells into bins across pseudotime and return expression
profile
}
\examples{
load(system.file("extdata", "test_seurat.rda", package = "IReNA"))
monocle_object = get_pseudotime(test_seurat)
seurat_object_pseudotime=add_pseudotime(seurat_object = test_seurat,monocle_object = monocle_object)
get_SmoothByBin_PseudotimeExp(seurat_object_pseudotime, Bin = 50, FcType = "Q95")
}
