% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth.R
\name{get_SmoothByBin_PseudotimeExp}
\alias{get_SmoothByBin_PseudotimeExp}
\title{get somoothbybin}
\usage{
get_SmoothByBin_PseudotimeExp(
  seurat_object,
  FC = TRUE,
  Bin = 50,
  FcType = "Q95"
)
}
\arguments{
\item{seurat_object}{seurat object, where meta.data should contain Pseudotime}

\item{FC}{logic, indicating whether to add FoldChangeQ95 to the first column, default is TRUE}

\item{Bin}{numeric, indicating the numbers of bin which divide the pseudotime, default is 50}

\item{FcType}{'Q95' or 'Q90', FoldChange threshold, default is 'Q95'}
}
\value{
return a expression profile
}
\description{
Divide cells into bins across pseudotime and return expression profile
}
\examples{
get_SmoothByBin_PseudotimeExp(seurat_object_pseudotime, Spec1 = 'Mm', Bin = 50, FcType = "Q95")
}
