% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_visualization.R
\name{initiate_cy}
\alias{initiate_cy}
\title{initiating the cytoscape}
\usage{
initiate_cy(TFs_list, colour = NULL, type = "TF", layout1 = "degree-circle")
}
\arguments{
\item{TFs_list}{TFs_list generated by \code{\link{network_analysis}}}

\item{colour}{vector, indicating colors for border of each node}

\item{type}{network type, 'TF' indicate TFs network, 'module' indicate intramodular network}

\item{layout1}{the layout to display the network, options: "degree-circle", "attributes-layout", "kamada-kawai", "force-directed", "cose", "hierarchical", "attribute-circle", "stacked-node-layout", "circular", "grid"}
}
\value{
Cytoscape visualization
}
\description{
Initiate Cytoscape to visualize regulatory network
}
\examples{
\dontrun{
initiate_cy(TFs_list, layout1='degree-circle', type='TF')
initiate_cy(TFs_list, layout1='grid', type='module')
}
}
