% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{merge_extent_footprints}
\alias{merge_extent_footprints}
\title{Inner function to merge and extent footprints}
\usage{
merge_extent_footprints(file1, motif)
}
\arguments{
\item{file1}{footprints}

\item{motif}{motif file}
}
\description{
Inner function to merge and extent footprints
}
