% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATAC-seq.R
\name{overlap_footprints_peaks}
\alias{overlap_footprints_peaks}
\title{Overlap differential peaks and motif footprints}
\usage{
overlap_footprints_peaks(footprints, peak_bed)
}
\arguments{
\item{footprints}{footprints file, generated by combine_footprints()}

\item{peak_bed}{bed format of peak, first column should be 'chr', second
column should be 'start', third column should be 'end'}
}
\value{
return data.frame contain footprints that overlap peaks
}
\description{
Because the running time of this function is too long, it's highly
recommanded to use bedtools to implement this function
}
\examples{
load(system.file("extdata", "test_peak.rda", package = "IReNA"))
load(system.file("extdata", "combined.rda", package = "IReNA"))
peak_bed <- get_bed(test_peak)
overlapped <- overlap_footprints_peaks(combined, peak_bed)
}
