% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_visualization.R
\name{plot_network}
\alias{plot_network}
\title{Display the network through igraph package}
\usage{
plot_network(
  network,
  layout,
  cols = NULL,
  cex = NULL,
  type = "TF",
  legend = TRUE
)
}
\arguments{
\item{network}{data.frame, indicating network file}

\item{layout}{the layout to display the network, options: 'grid','sphere','circle','random'}

\item{cols}{vector, indicating cols of each group}

\item{cex}{numeric, indicating the font size of node labels}

\item{type}{network type, 'TF' indicate TFs network, 'module' indicate intramodular network}

\item{legend}{TRUE or FALSE, whether to show the legend}
}
\value{
figure
}
\description{
Display the network through igraph package
}
\examples{
plot_network(tf_network_cor0.6_thr10, layout='circle', type='TF')

}
