% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Footprints_FOS.R
\name{Trans_WigToMultirows}
\alias{Trans_WigToMultirows}
\title{Inner function}
\usage{
Trans_WigToMultirows(Wig_list)
}
\arguments{
\item{Wig_list}{list that each element is wig dataframe}
}
\description{
Inner function
}
