% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wig_track.R
\name{wig_track}
\alias{wig_track}
\title{wig track}
\usage{
wig_track(bamfilepath, bedfile, index = FALSE)
}
\arguments{
\item{bamfilepath}{character, indicating the path of bam file}

\item{bedfile}{bed file of footprints, generated by get_peaks_genes()}

\item{index}{logical, indicating whether to index the bam file. If bam
file do not have a matched bai file, this parameter should be TRUE}
}
\value{
return formated wig file
}
\description{
Function to get cuts of each position in footprints
}
\examples{
load(system.file("extdata", "list1.rda", package = "IReNA"))
load(system.file("extdata", "test_clustering.rda", package = "IReNA"))
bamfilepath1<-'mmATACCtrW00R1_CuFiQ10No_sorted.bam'
bamfilepath2<-'mmATACCtrW00R2_CuFiQ10No_sorted.bam'
test_clustering <- add_ENSID(test_clustering, Spec1 = "Hs")
list2<-get_related_peaks(list1,test_clustering)
#cuts1<-wig_track(bamfilepath = bamfilepath1,bedfile = list2[[2]])
#cuts2<-wig_track(bamfilepath = bamfilepath2,bedfile = list2[[2]])
#wig_list<-list(cuts1,cuts2)
}
