% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wig_track.R
\name{wig_track}
\alias{wig_track}
\title{Function to get cuts of each position in footprints}
\usage{
wig_track(bamfilepath, bedfile, index = FALSE)
}
\arguments{
\item{bamfilepath}{character, indicating the path of bam file}

\item{bedfile}{bed file of footprints, generated by get_peaks_genes()}

\item{index}{logical, whether to index the bam file. If bam file do not have a matched bai file, this parameter should be TRUE}
}
\value{
return formated wig file
}
\description{
Function to get cuts of each position in footprints
}
\examples{
bamfilepath1<-'mmATACCtrW00R1_CuFiQ10No_sorted.bam'
bamfilepath2<-'mmATACCtrW00R2_CuFiQ10No_sorted.bam'
list2<-get_peaks_genes(list1,MmscRNA_PHx_Exp_NewF)
#cuts1<-wig_track(bamfilepath1 = bamfilepath,bedfile = list2[[2]])
#cuts2<-wig_track(bamfilepath2 = bamfilepath,bedfile = list2[[2]])
#wig_list<-list(cuts1,cuts2)
}
